SessionCache: new implementation
The previous cache wasn't yet complete, but it would eventually require
maintaining two underlying data structures: a Map<String, Session> and a
PriorityQueue<Session>.

This change replaces that implementation with a single underlying data
structure: a hash table with no resizing and replace-the-previous-entry
collision handling.  The cache pruning is dealt with automatically
because of the more frequent collisions resulting from filling the
table.

The upside: (complete,) simpler and faster implementation.

The downside: there is a small but not negligible probability that a
pair or simultaneous users have colliding sessions.  To minimize the
frequency and impact of this, it may be necessary to carefully tune the
hash function.