Fix: upvote/downvote onFail ui behavior improved, but still dumb
We most likely need to check the current server status, because we
can't be sure if the server has processed the request or not.  For
example, it might have failed on the reputation on notification systems,
but has already saved the upvote.

Also: we need to assert correct reputation updates, particularly if they
are synchronous (of it the users expects them to be).